confusing bs

+++ src/main/java/io/disc/DiscLoader/objects/gateway/RoleJSON.java
@@ -3,9 +3,9 @@ package io.disc.DiscLoader.objects.gateway;
 public class RoleJSON {
 	public String id;
 	public String name;
-	public int permissions;
-	public int color;
-	public int position;
+	public double permissions;
+	public double color;
+	public double position;
 	public boolean hoist;
 	public boolean managed;
 	public boolean mentionable;
